Добавил доп сид данные + доделал пагинацию студентов
В пагинации некорректно работали страницы, теперь всё ровно

+++ Finally/School/School.Core/Models/Pages/PageList.cs
@@ -9,12 +9,12 @@ namespace School.Core.Models.Pages
         {
             CurrentPage = options.CurrentPage;
             PageSize = options.PageSize;
-            TotalPages = query.Count() / PageSize;
+            TotalPages = (double)query.Count() / (double)PageSize;
             AddRange(query.Skip((CurrentPage - 1) * PageSize).Take(PageSize));
         }
         public int CurrentPage { get; set; }
         public int PageSize { get; set; }
-        public int TotalPages { get; set; }
+        public double TotalPages { get; set; }
         public bool HasPreviosPage => CurrentPage > 1;
         public bool HasNextPage => CurrentPage < TotalPages;
     }
